change: turn `Account` into an aggregate root
