+++ src/modules/identity/errors/account-not-found.error.ts
@@ -1,5 +1,5 @@
 import { ResourceNotFoundError } from "@/common/errors/resource-not-found.error";
-import { Account } from "@/modules/identity/entities/account.entity";
+import { Account } from "@/modules/identity/entities/account.aggregate";
 
 export class AccountNotFoundError extends ResourceNotFoundError {
   public constructor() {
